agent-python/requirements.txt
altgraph==0.17.4
bcrypt==4.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
dotenv==0.9.9
grpc-stubs==1.53.0.5
grpcio==1.71.0
grpcio-tools==1.71.0
idna==3.10
mypy==1.15.0
mypy-extensions==1.0.0
mypy-protobuf==3.6.0
packaging==24.2
paramiko==3.5.1
protobuf==5.29.4
pycparser==2.22
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.2
PyNaCl==1.5.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
types-paramiko==3.5.0.20240928
types-protobuf==5.29.1.20250403
types-python-crontab==3.2.0.20240703
types-PyYAML==6.0.12.20250402
types-requests==2.32.0.20250328
typing_extensions==4.13.2
urllib3==2.4.0
